requirements.txt
vosk==0.3.45
pyaudio==0.2.14
requests==2.32.3
python-dotenv==1.0.1
ollama==0.3.1

